Metrics lib

+++ gydam-db/index.js
@@ -5,6 +5,7 @@ const setupAgentModel = require('./models/agent')
 const setupMetricModel = require('./models/metric')
 const defaults = require('defaults')
 const setupAgent = require('./lib/agent')
+const setupMetric = require('./lib/metric')
 
 module.exports = async function (config) {
   config = defaults(config, {
@@ -28,12 +29,10 @@ module.exports = async function (config) {
 
   await sequelize.authenticate()
 
-  if (config.setup) {
-    await sequelize.sync({ force: true })
-  }
+  if (config.setup) await sequelize.sync({ force: true })
 
   const Agent = setupAgent(AgentModel)
-  const Metric = {}
+  const Metric = setupMetric(MetricModel, AgentModel)
 
   return {
     Agent,
